requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.14.0
psycopg2==2.9.1 
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.2
uritemplate==4.1.1